worker: split the hub session loop out of worker.rs
No behavior change.
